[API-633] Manage Bunyan's Stream Manually
This is a bit of a cleaner approach -- don't give Bunyan any file streams to handle. Then, expose a really simple method, "cleanStream" that will do a delayed close of the request stream. We'll use this at the end of the restify logger, plus we need to use it elsewhere.

https://jira.appcelerator.org/browse/API-633